beginning spill out

+++ L2/src/interference.cpp
@@ -8,24 +8,59 @@
 
 namespace L2 {
 
+	bool printI = false;
+
 	InterferenceGraph::InterferenceGraph(Function* f) {
+		this->funct = f;
 		// create item vector
-		for (auto i : f->instructions) {
-			for (auto var : i->gen)
-				if (!(var->is_in(this->items))) this->items.push_back(var);
-			for (auto var : i->kill)
-				if (!(var->is_in(this->items))) this->items.push_back(var);
+		if (printI) std::cout << "creating item and reduce" << std::endl;
+		if (printI) std::cout << "adding gprs" << std::endl;
+		for (auto var : this->gprs) {
+			if (!(var->is_in(f->items))) {
+				f->items.push_back(var);
+				f->reduce_v[var] = var;
+			} else f->reduce_v[var] = var->get_dup(f->items);
 		}
-		for (auto var : this->gprs)
-			if (!(var->is_in(this->items))) this->items.push_back(var);
+		if (printI) std::cout << "entering instructions: " << f->instructions.size() << std::endl;
+		for (auto i : f->instructions) {
+			if (printI) std::cout << "adding gens" << std::endl;
+			for (auto var : i->gen) {
+				if (!(var->is_in(f->items))) {
+					f->items.push_back(var);
+					f->reduce_v[var] = var;
+				} else f->reduce_v[var] = var->get_dup(f->items);
+			}
+			if (printI) std::cout << "adding kills" << std::endl;
+			for (auto var : i->kill) {
+				if (!(var->is_in(f->items))) {
+					f->items.push_back(var);
+					f->reduce_v[var] = var;
+				} else f->reduce_v[var] = var->get_dup(f->items);
+			}
+			if (printI) std::cout << "adding ins" << std::endl;
+			for (auto var : i->in) {
+				if (!(var->is_in(f->items))) {
+					f->items.push_back(var);
+					f->reduce_v[var] = var;
+				} else f->reduce_v[var] = var->get_dup(f->items);
+			}
+			if (printI) std::cout << "adding outs" << std::endl;
+			for (auto var : i->out) {
+				if (!(var->is_in(f->items))) {
+					f->items.push_back(var);
+					f->reduce_v[var] = var;
+				} else f->reduce_v[var] = var->get_dup(f->items);
+			}
+		}
+		if (printI) std::cout << "done" << std::endl;
 		// create the index mapping
 		int index = 0;
-		for (auto var : this->items) {
+		for (auto var : f->items) {
 			this->indices[var] = index;
 			index++;
 		}
-		std::cout << "num indices: " << this->indices.size() << std::endl;
-		std::cout << "num items: " << this->items.size() << std::endl;
+		if (printI) std::cout << "num indices: " << this->indices.size() << std::endl;
+		if (printI) std::cout << "num items: " << f->items.size() << std::endl;
 		// create the matrix
 		for (int i=0; i<(index*index); i++)
 			this->adj_matrix.push_back(false);
@@ -33,15 +68,11 @@ namespace L2 {
 		// begin interference analysis
 
 		// connect all gp registers
-		for (auto reg1 : this->items) {
-			if (reg1->is_in(this->gprs)) {
-				for (auto reg2 : this->items) {
-					if (reg2->is_in(this->gprs)) {
-						if (reg1 != reg2) {
-							std::cout << "connect gprs" << std::endl;
-							this->add_edge(reg1, reg2);
-						}
-					}
+		for (auto reg1 : this->gprs) {
+			for (auto reg2 : this->gprs) {
+				if (f->reduce_v[reg1] != f->reduce_v[reg2]) {
+					if (printI) std::cout << "connect gprs" << std::endl;
+					this->add_edge(f->reduce_v[reg1], f->reduce_v[reg2]);
 				}
 			}
 		}
@@ -49,17 +80,17 @@ namespace L2 {
 		for (auto instr : f->instructions) {
 			for (auto var1 : instr->in) {
 				for (auto var2 : instr->in) {
-					if (var1 != var2) {
-						std::cout << "connect ins" << std::endl;
-						this->add_edge(var1, var2);
+					if (f->reduce_v[var1] != f->reduce_v[var2]) {
+						if (printI) std::cout << "connect ins" << std::endl;
+						this->add_edge(f->reduce_v[var1], f->reduce_v[var2]);
 					}
 				}
 			}
 			for (auto var1 : instr->out) {
 				for (auto var2 : instr->out) {
-					if (var1 != var2) {
-						std::cout << "connect outs" << std::endl;
-						this->add_edge(var1, var2);
+					if (f->reduce_v[var1] != f->reduce_v[var2]) {
+						if (printI) std::cout << "connect outs" << std::endl;
+						this->add_edge(f->reduce_v[var1], f->reduce_v[var2]);
 					}
 				}
 			}
@@ -68,9 +99,9 @@ namespace L2 {
 		for (auto instr : f->instructions) {
 			for (auto var1 : instr->kill) {
 				for (auto var2 : instr->out) {
-					if (var1 != var2) {
-						std::cout << "connect kill out" << std::endl;
-						this->add_edge(var1, var2);
+					if (f->reduce_v[var1] != f->reduce_v[var2]) {
+						if (printI) std::cout << "connect kill out" << std::endl;
+						this->add_edge(f->reduce_v[var1], f->reduce_v[var2]);
 					}
 				}
 			}
@@ -78,11 +109,10 @@ namespace L2 {
 		// handle target language constraints
 		for (auto instr : f->instructions) {
 			if (dynamic_cast<Instruction_sop_lsh*> (instr) || dynamic_cast<Instruction_sop_rsh*> (instr)) {
-				//instr->gen[0]
-				for (auto reg : this->items) {
-					if (reg->is_in(this->gprs) && reg->name.compare("rcx") != 0) {
-						std::cout << "rcx shift" << std::endl;
-						this->add_edge(instr->gen[0], reg);
+				for (auto reg : this->gprs) {
+					if (reg->name.compare("rcx") != 0) {
+						if (printI) std::cout << "rcx shift" << std::endl;
+						this->add_edge(f->reduce_v[instr->gen[0]], f->reduce_v[reg]);
 					}
 				}
 			}
@@ -92,20 +122,22 @@ namespace L2 {
 	void InterferenceGraph::add_edge(Variable* start, Variable* end) {
 		this->adj_matrix[this->indices[start]*this->indices.size() + this->indices[end]] = true;
 		this->adj_matrix[this->indices[end]*this->indices.size() + this->indices[start]] = true;
-		std::cout << "num indices: " << this->indices.size() << std::endl;
-		std::cout << "num items: " << this->items.size() << std::endl;
+		if (printI) std::cout << "num indices: " << this->indices.size() << std::endl;
+		if (printI) std::cout << "num items: " << this->funct->items.size() << std::endl;
 	}
 
 	void generate_interference_output(InterferenceGraph* ig) {
-		std::cout << "num indices: " << ig->indices.size() << std::endl;
-		std::cout << "num items: " << ig->items.size() << std::endl;
+		//if (printI) std::cout << "num indices: " << ig->indices.size() << std::endl;
+		//if (printI) std::cout << "num items: " << ig->items.size() << std::endl;
 		for (int i=0; i<ig->indices.size(); i++) {
-			std::cout << "printing name: ";
-			std::cout << ig->items[i]->name << std::endl;
+			//if (printI) std::cout << "printing name: ";
+			//if (printI) std::cout << ig->items[i]->name << std::endl;
+			std::cout << ig->funct->items[i]->name;
 			for (int j=0; j<ig->indices.size(); j++) {
-				std::cout << "potential interf: " << ig->items[j]->name << std::endl;
+				//if (printI) std::cout << "potential interf: " << ig->items[j]->name << std::endl;
 				if (i!=j && ig->adj_matrix[ig->indices.size()*i + j])
-					std::cout << "yes! " << ig->items[j]->name << std::endl;
+					//if (printI) std::cout << "yes! " << ig->items[j]->name << std::endl;
+					std::cout << " " << ig->funct->items[j]->name;
 			}
 			std::cout << "\n";
 		}
